MA-9677 Fix new gralloc crash issue on imx8qm
Use intptr_t to suitable both 32bit and 64bit platform, intptr_t is defined
as int in 32bi and log int in 64bit, that can make sure no precision lost
when do convert with pointer.

crash log:
F libc    : Fatal signal 11 (SIGSEGV), code 1, fault addr 0xffffffc6e4606c in tid 3437 (Binder:3371_3)
F DEBUG   : Build fingerprint: 'Freescale/arm2_8qm/arm2_8qm:7.1.2/1.0.0-beta-rc7/20161226:eng/release-keys'
F DEBUG   : Revision: '0'
F DEBUG   : ABI: 'arm64'
F DEBUG   : pid: 3371, tid: 3437, name: Binder:3371_3  >>> /system/bin/surfaceflinger <<<
F DEBUG   : signal 11 (SIGSEGV), code 1 (SEGV_MAPERR), fault addr 0xffffffc6e4606c
F DEBUG   :     x0   ffffffffc6e4606c  x1   0000e869c6a7d6c0  x2   0000e869d1bb6a14  x3   0000000000000002
F DEBUG   :     x4   0000000030000b00  x5   0000000000000002  x6   0000000030000300  x7   0064006f00720050
F DEBUG   :     x8   0000e869c7108450  x9   0000000003141592  x10  0000000000000001  x11  0000000000000000
F DEBUG   :     x12  0000e869c6a97270  x13  0072005000720065  x14  006300750064006f  x15  0000000000000000
F DEBUG   :     x16  0000e869d1b359d0  x17  0000e869d16e50b0  x18  0000000000017000  x19  ffffffffc6e46060
F DEBUG   :     x20  ffffffffc6e4606c  x21  b06cc4ea0625accf  x22  0000e869d199333c  x23  0000e869d1993030
F DEBUG   :     x24  0000000000000780  x25  00000000ffffffff  x26  0000e869c6a9c070  x27  0000e869d199333c
F DEBUG   :     x28  0000e869c6a9c070  x29  0000e869c71079f0  x30  0000e869d1b2e624
F DEBUG   :     sp   0000e869c71079d0  pc   0000e869d16e50b0  pstate 0000000060000000
F DEBUG   :
F DEBUG   : backtrace:
F DEBUG   :     #00 pc 00000000000690b0  /system/lib64/libc.so (pthread_mutex_lock)
F DEBUG   :     #01 pc 000000000000c620  /system/lib64/libfsldisplay.so (_ZN3fsl12MemoryShadow6decRefEv+28)
F DEBUG   :     #02 pc 000000000000c820  /system/lib64/libfsldisplay.so (_ZN3fsl13MemoryManager13releaseMemoryEPNS_6MemoryE+60)
F DEBUG   :     #03 pc 0000000000001a48  /system/lib64/hw/gralloc.imx8.so
F DEBUG   :     #04 pc 000000000001137c  /system/lib64/libui.so (_ZN7android22GraphicBufferAllocator4freeEPK13native_handle+92)
F DEBUG   :     #05 pc 000000000000fe10  /system/lib64/libui.so (_ZN7android13GraphicBuffer11free_handleEv+108)
F DEBUG   :     #06 pc 000000000000fe88  /system/lib64/libui.so (_ZN7android13GraphicBufferD0Ev+44)
F DEBUG   :     #07 pc 000000000000e378  /system/lib64/libutils.so (_ZNK7android7RefBase9decStrongEPKv+92)
F DEBUG   :     #08 pc 00000000000626cc  /system/lib64/libgui.so (_ZN7android19BufferQueueProducer13dequeueBufferEPiPNS_2spINS_5FenceEEE)
F DEBUG   :     #09 pc 0000000000070054  /system/lib64/libgui.so (_ZN7android23BnGraphicBufferProducer10onTransactEjRKNS_6ParcelEPS1_j+3)
F DEBUG   :     #10 pc 0000000000049e2c  /system/lib64/libbinder.so (_ZN7android7BBinder8transactEjRKNS_6ParcelEPS1_j+132)
F DEBUG   :     #11 pc 0000000000055bec  /system[   38.706555] Binder:3371_3[3437]: unhandled level 0 translation fault (11) at 0xffffff4

Reviewed-on: http://androidsource.ap.freescale.net/project/2169

+++ display/display/GPUManager.cpp
@@ -111,7 +111,7 @@ int GPUManager::allocMemory(MemoryDesc& desc, Memory** out)
     }
 
     MemoryShadow* shadow = new GPUShadow(memory, true, mAlloc, mModule);
-    memory->shadow = (int)(intptr_t)shadow;
+    memory->shadow = (intptr_t)shadow;
     *out = memory;
 
     return 0;
@@ -138,7 +138,7 @@ int GPUManager::retainMemory(Memory* handle)
         }
 
         shadow = new GPUShadow(handle, false, mAlloc, mModule);
-        handle->shadow = (int)(intptr_t)shadow;
+        handle->shadow = (intptr_t)shadow;
         handle->pid = getpid();
     }
     else {
